feat(security): implement RBAC with UserRole, Permission, and warehouse isolation
- Add UserRole enum (admin, accountant, storeManager, viewer)
- Add Permission bitmask class for granular access control
- Extend BasirUser with role, permissions, and warehouseId fields
- Add getCurrentUser() with full RBAC data retrieval
- Add hasPermission() and hasAccessToWarehouse() methods
- Update createAccount() to support role and warehouse assignment
- Fix invoice print service localization
- Fix business module lint issues
- Modernize auth_service.dart with cleaner implementation

+++ lib/features/dashboard/domain/entities/dashboard_data.dart
@@ -1,4 +1,5 @@
 import 'package:basir_app/features/invoices/domain/entities/invoice.dart';
+import 'package:decimal/decimal.dart';
 
 /// بيانات لوحة التحكم (Dashboard Data)
 ///
@@ -33,16 +34,16 @@ class DashboardData {
   final int pendingInvoices;
 
   /// إجمالي المبيعات (جميع الفواتير غير الملغاة)
-  final double totalSales;
+  final Decimal totalSales;
 
   /// المبالغ المحصلة (المدفوعة فعلياً)
-  final double paidRevenue;
+  final Decimal paidRevenue;
 
   /// المبالغ المعلقة (صادرة وغير مدفوعة وليست متأخرة)
-  final double pendingRevenue;
+  final Decimal pendingRevenue;
 
   /// المبالغ المتأخرة (تجاوزت تاريخ الاستحقاق)
-  final double overdueRevenue;
+  final Decimal overdueRevenue;
 
   /// عدد العملاء النشطين
   final int activeCustomersCount;
